requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
anyio==4.4.0
apache-airflow==2.10.0
apache-airflow-providers-common-compat==1.1.0
apache-airflow-providers-common-io==1.4.0
apache-airflow-providers-common-sql==1.15.0
apache-airflow-providers-fab==1.2.2
apache-airflow-providers-ftp==3.10.1
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.8.2
apispec==6.6.1
argcomplete==3.5.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
blinker==1.8.2
cachelib==0.9.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==6.8.2
ConfigUpdater==3.2
confluent-kafka==2.5.0
connexion==2.14.2
cron-descriptor==1.4.3
croniter==3.0.3
cryptography==43.0.0
Deprecated==1.2.14
dill==0.3.8
dnspython==2.6.1
docutils==0.21.2
email_validator==2.2.0
exceptiongroup==1.2.2
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.6.1
google-re2==1.1.20240702
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.5
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.3
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.4.0
multidict==6.0.5
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
ordered-set==4.1.0
packaging==24.1
pathspec==0.12.1
pendulum==3.0.0
pluggy==1.5.0
prison==0.2.1
protobuf==4.25.4
psutil==6.0.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.5.2
rpds-py==0.20.0
setproctitle==1.3.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.53
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.1
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.15.0
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
urllib3==2.2.2
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.20.0
